# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "spassu",
        "version": "1"
    },
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "subjects"
                ],
                "summary": "List all Subjects",
                "operationId": "26bbe2e175802183241930c0ebb0c6a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the subject"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Subject"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "subjects"
                ],
                "summary": "Add new Subject",
                "operationId": "c18cbfbfacfa1b499acb40d9901bf383",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddSubjectDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "$ref": "#/components/schemas/Subject"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "subjects"
                ],
                "summary": "Delete Subject",
                "operationId": "18b76cb5cb2d4b0366371f44fd983a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the subject"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/{id}": {
            "get": {
                "tags": [
                    "subjects"
                ],
                "summary": "Get Subject",
                "operationId": "4cdc74f682711aa11a8d340c5f264c5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the subject"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "$ref": "#/components/schemas/Subject"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "subjects"
                ],
                "summary": "Update Subject",
                "operationId": "baf660d47dc7cadf37a836d0e51b6a24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the subject"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "$ref": "#/components/schemas/Subject"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AddSubjectDTO": {
                "title": "AddSubjectDTO",
                "properties": {
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Subject": {
                "title": "Subject",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}